----- fx.js ----- added: $.fn.fadeTo Also changed ' fx.Opacity ' function so $.fn.fadeTo works. For this i changed one line, and one parameter in $.fx.Opactity: "o.io = o.now = 1;" became: "o.io = o.now = (sv || o.cur());"
e.g. if there is a start value use it. Else, the current value is used. All calls to "$.fx.Opactiy" which where present in the code have now "1" as third parameter so the code doesn't break.

---------
jquery.js
---------
Changed line 401 to fix ' name="id" ' bug:
'#': "a.id == m[2]"
to:
'#': "a.getAttribute('id') == m[2]"


git-svn-id: http://jqueryjs.googlecode.com/svn/jquery@41 c715fcbe-d12f-0410-84c4-316a508785bb